RD-5177 Move Getting Started schemas to S3
